Make testcomposer client implement the Job.Writer interface
Uploads go to test-composer, but the upload logic was implemented by the
resto client.
